+++ src/components/exhibition/index.ts
@@ -1,12 +1,12 @@
 // 예정, 종료 페이지
-export { default as ExhibitionEnded } from './exhibitionEnded';
-export { default as ExhibitionUpcoming } from './exhibitionUpcoming';
+export { default as ExhibitionEnded } from './ExhibitionEnded';
+export { default as ExhibitionUpcoming } from './ExhibitionUpcoming';
 
 // actionBar
-export { default as ExhibitionActionBar } from './exhibitionActionBar';
+export { default as ExhibitionActionBar } from './ExhibitionActionBar';
 
 // review
-export { default as ReviewSection } from './reviewSection';
+export { default as ReviewSection } from './ReviewSection';
 
 // 작품 dialog
-export { default as WorkDialog } from './workDialog';
+export { default as WorkDialog } from './WorkDialog';
