+++ src/components/exhibition/manage/AddWorkDialog.tsx
@@ -1,8 +1,9 @@
 'use client';
 
-import { Pencil } from 'lucide-react';
+import { Plus } from 'lucide-react';
+import { cn } from '@/lib/utils';
 import { Button } from '@/components/ui/button';
-import WorkFormBox from './workFormBox';
+import WorkFormBox from './WorkFormBox';
 import {
   Dialog,
   DialogClose,
@@ -11,32 +12,31 @@ import {
   DialogTitle,
   DialogTrigger,
 } from '@/components/ui/dialog';
-import ImageUploadBox from './imageUploadBox';
-import { useState } from 'react';
-import { useParams, useRouter } from 'next/navigation';
+import ImageUploadBox from './ImageUploadBox';
 import { Controller, useForm } from 'react-hook-form';
-import { putArtWorkByArtWorkId } from '@/service/artworks';
-import { ArtworkFormUi } from '@/components/exhibition/manage/addWorkDialog';
-import { ArtworkWithEmail } from '@/app/(exhibitions)/exhibitions/[id]/manage/page';
+import { useParams, useRouter } from 'next/navigation';
+import { postArtWorksByExhibitionId } from '@/service/artworks';
+import { useState } from 'react';
 
-export interface Work {
-  id: string;
+interface AddWorkDialogProps {
+  triggerLabel?: string;
+  triggerClassName?: string;
+}
+export type ArtworkFormUi = {
+  artist_email: string | null;
   title: string;
   artist_name: string;
-  image_url: string;
-  artist_email?: string;
-  description?: string;
-  artist_id?: string;
-}
-
-interface EditWorkDialogProps {
-  work: ArtworkWithEmail;
-}
+  description: string | null;
+  image_url: string | File;
+};
 
 const fieldClass =
   'text-secondary placeholder:text-secondary/40 w-full rounded-xl border border-gray-200 bg-surface px-4 py-3 text-sm outline-none focus:border-[#F5A623] focus:bg-white';
 
-export default function EditWorkDialog({ work }: EditWorkDialogProps) {
+export default function AddWorkDialog({
+  triggerLabel = '작품 추가',
+  triggerClassName,
+}: AddWorkDialogProps) {
   const { id } = useParams<{ id: string }>();
   const [open, setOpen] = useState(false);
   const router = useRouter();
@@ -49,11 +49,11 @@ export default function EditWorkDialog({ work }: EditWorkDialogProps) {
   } = useForm<ArtworkFormUi>({
     mode: 'onChange',
     defaultValues: {
-      artist_email: work?.artist_email || null,
-      title: work.title,
-      artist_name: work.artist_name,
-      description: work.description,
-      image_url: work.image_url,
+      artist_email: null,
+      title: '',
+      artist_name: '',
+      description: null,
+      image_url: '',
     },
   });
   const submitHandler = async (e: ArtworkFormUi) => {
@@ -69,13 +69,14 @@ export default function EditWorkDialog({ work }: EditWorkDialogProps) {
     formData.append('image_url', e.image_url);
     console.log(e);
     try {
-      const editArtWorkId = await putArtWorkByArtWorkId(id, work.id, formData);
-      console.log(editArtWorkId);
+      const createdId = await postArtWorksByExhibitionId(id, formData);
+      console.log(createdId);
       setOpen(false);
       router.refresh();
     } catch (e) {
       console.log(e);
       if (e instanceof Error && e.message === 'profile not found') {
+        console.log('dnsjehsndjwndj');
         setError('artist_email', {
           type: 'server',
           message: '가입된 프로필이 없습니다',
@@ -83,25 +84,20 @@ export default function EditWorkDialog({ work }: EditWorkDialogProps) {
       }
     }
   };
+
   return (
     <Dialog open={open} onOpenChange={setOpen}>
       <DialogTrigger
-        render={
-          <Button
-            size="sm"
-            variant="surface"
-            className="hover:text-primary flex-1 rounded-lg"
-          />
-        }
+        render={<Button className={cn('rounded-xl', triggerClassName)} />}
       >
-        <Pencil className="h-3.5 w-3.5" />
-        수정
+        <Plus className="h-4 w-4" />
+        {triggerLabel}
       </DialogTrigger>
 
       <DialogContent className="max-h-[90vh] w-[calc(100%-2rem)] max-w-xl overflow-y-auto sm:max-w-lg">
         <DialogHeader>
           <DialogTitle className="text-secondary text-lg font-bold">
-            작품 수정
+            작품 등록
           </DialogTitle>
         </DialogHeader>
 
@@ -127,6 +123,7 @@ export default function EditWorkDialog({ work }: EditWorkDialogProps) {
               <input
                 {...register('title', { required: true })}
                 type="text"
+                placeholder="작품 제목을 입력하세요"
                 className={fieldClass}
               />
             </WorkFormBox>
@@ -135,18 +132,12 @@ export default function EditWorkDialog({ work }: EditWorkDialogProps) {
               <input
                 {...register('artist_name', { required: true })}
                 type="text"
+                placeholder="학생 이름"
                 className={fieldClass}
               />
             </WorkFormBox>
 
-            <WorkFormBox
-              label={
-                <>
-                  작가 이메일{' '}
-                  <span className="text-secondary/50 font-normal">(선택)</span>
-                </>
-              }
-            >
+            <WorkFormBox label="작가 이메일">
               <input
                 {...register('artist_email', { required: false })}
                 type="email"
@@ -175,11 +166,11 @@ export default function EditWorkDialog({ work }: EditWorkDialogProps) {
 
           <div className="flex gap-2 pt-2">
             <Button
-              disabled={!isValid || isSubmitting}
               type={'submit'}
+              disabled={!isValid || isSubmitting}
               className="flex-1 rounded-xl py-6"
             >
-              수정하기
+              {isSubmitting ? '등록중' : '등록하기'}
             </Button>
             <DialogClose
               render={
