+++ src/components/exhibition/manage/index.ts
@@ -1,6 +1,6 @@
 // 작품 추가, 수정버튼 dialog
-export { default as EditWorkDialog } from './editWorkDialog';
-export { default as AddWorkDialog } from './addWorkDialog';
+export { default as EditWorkDialog } from './EditWorkDialog';
+export { default as AddWorkDialog } from './AddWorkDialog';
 
 // 전시회 종료버튼, 삭제버튼 alert dialog
-export { default as ManageAlertDialog } from './manageAlertDialog';
+export { default as ManageAlertDialog } from './ManageAlertDialog';
